Update FATIP-0 to match changes to FATIP-100
- Removed "extends FATIP-100" from the Token ID field since the Token ID
is not really extending FATIP-100 but is used in the derivation of Chain
IDs defined by FATIP-100. The Issuance and Transaction Chain IDs do
still extend FATIP-100.
- Improved Token ID description and mentioned the fact that two Issuers
can use the same Token ID.
- Changed link to Factom Documentation to the Github page so that we can
link to the specifc section as opposed to the entire Factom.com doc
page.
- Reiterated the External IDs used for the Issuance and Transaction
Chain IDs and added examples. This is redundant because it is described
in FATIP-100 but I believe that it is helpful to have in both places.  -
Added detail that the officially valid Issuance Entry does NOT have to
be the very first entry in the chain. Simply the first well-formed
validly signed Issuance Entry in the chain. This prevents Token IDs from
being DOS'd since only the entity that controls the Issuer Identity
pointed to by the Chain ID in the Chain's second External ID may create
a validly signed Issuance Entry.
- Expanded the validation column for the issuer field in the Issuance
Entry field table.