Add MIMO tx signing with reduced db sig lookups
This adds transaction signing for multi-input transactions using Factoid
Address RCD tx signing structure. However unlike the binary Factoid
Transaction datastructure, RCDs and signatures are stored in the
External IDs of the transaction entry.

Additionally this commit greatly improves the specificity of transaction
validation requirements and introduces a convention for numbering
validation requirements. Validation requirements are roughly in order of
computational and then programmatic complexity, this way implementations
that follow the ordering can rule out invalid transactions as
efficiently as possible.

The computation of state is now separated from the description of the
validity rules.

In order to reduce the requirement to search all previous signatures on
valid transactions, the field `declaredHeight` scopes the search to
transactions with the same `declaredHeight`. Validity rule N.2.1 ensures
that implementations only need to store the current and the previously
scanned block of valid transactions in memory. The reason that an
overlap is allowed is because Factom does not guarantee that an entry
will make it into the current block if it is revealed past the block
cutoff time.

TODO: Update the Implementation sections according to these changes.